Remove fallback url origin in favor of invariant
